+++ tests/dummy/app/JsonApi/Images/Adapter.php
@@ -1,0 +1,63 @@
+<?php
+/**
+ * Copyright 2020 Cloud Creativity Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace DummyApp\JsonApi\Images;
+
+use CloudCreativity\LaravelJsonApi\Eloquent\AbstractAdapter;
+use CloudCreativity\LaravelJsonApi\Pagination\StandardStrategy;
+use DummyApp\Image;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Collection;
+
+class Adapter extends AbstractAdapter
+{
+
+    /**
+     * Mapping of JSON API attribute field names to model keys.
+     *
+     * @var array
+     */
+    protected $attributes = [];
+
+    /**
+     * Mapping of JSON API filter names to model scopes.
+     *
+     * @var array
+     */
+    protected $filterScopes = [];
+
+    /**
+     * Adapter constructor.
+     *
+     * @param StandardStrategy $paging
+     */
+    public function __construct(StandardStrategy $paging)
+    {
+        parent::__construct(new Image(), $paging);
+    }
+
+    /**
+     * @param Builder $query
+     * @param Collection $filters
+     * @return void
+     */
+    protected function filter($query, Collection $filters)
+    {
+        $this->filterWithScopes($query, $filters);
+    }
+
+}