+++ tests/dummy/app/JsonApi/Images/Schema.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace DummyApp\JsonApi\Images;
+
+use DummyApp\Image;
+use Neomerx\JsonApi\Schema\SchemaProvider;
+
+class Schema extends SchemaProvider
+{
+
+    /**
+     * @var string
+     */
+    protected $resourceType = 'images';
+
+    /**
+     * @param Image $resource
+     *      the domain record being serialized.
+     * @return string
+     */
+    public function getId($resource)
+    {
+        return (string) $resource->getRouteKey();
+    }
+
+    /**
+     * @param Image $resource
+     *      the domain record being serialized.
+     * @return array
+     */
+    public function getAttributes($resource)
+    {
+        return [
+            'created-at' => $resource->created_at->toAtomString(),
+            'updated-at' => $resource->updated_at->toAtomString(),
+            'url' => $resource->url,
+        ];
+    }
+}