+++ tests/dummy/database/migrations/2018_02_11_1648_create_tables.php
@@ -48,6 +48,13 @@ class CreateTables extends Migration
             $table->unsignedInteger('user_id')->nullable();
         });
 
+        Schema::create('images', function (Blueprint $table) {
+            $table->increments('id');
+            $table->timestamps();
+            $table->string('url');
+            $table->nullableMorphs('imageable');
+        });
+
         Schema::create('posts', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();